% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{arraydata}
\alias{arraydata}
\title{Example microarray for the study of Ezh2.}
\description{
Example microarray for the study of Ezh2.
}
\references{
\url{http://www.cell.com/cell-reports/abstract/S2211-1247(13)00007-7}
}
\author{
Bhupinder Pal, Toula Bouras, Wei Shi, Francois Vaillant, Julie M. Sheridan,
  Naiyang Fu, Kelsey Breslin, Kun Jiang, Matthew E. Ritchie, Matthew Young,
  Geoffrey J. Lindeman, Gordon K. Smyth, Jane E. Visvader
}
\keyword{microarray}
