% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CharToHexCol}
\alias{CharToHexCol}
\title{String to hex colour converter}
\usage{
CharToHexCol(x)
}
\arguments{
\item{x}{the string colour value(s) to be converted to hex values.}
}
\value{
hex codes for colours
}
\description{
Function to convert colour strings into hex codes
}

