% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glSearch.R
\name{glAutocomplete}
\alias{glAutocomplete}
\title{Create a jsinput object containing information for glimma function to 
generate an autocompletion field.}
\usage{
glAutocomplete(target, action)
}
\arguments{
\item{index}{the index of the chart from which data is obtained}

\item{idcol}{the column name of the data to use for input}
}
\value{
a jsinput object containing the data for creating an autocomplete field
}
\description{
Create a jsinput object containing information for glimma function to 
generate an autocompletion field.
}
\examples{

}

