% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glInputs.R
\name{glAutoinput}
\alias{glAutoinput}
\title{Helper function for writing input fields in interactive Glimma plots}
\usage{
glAutoinput(target, action, idval = "none", flag = "none")
}
\arguments{
\item{target}{the index of the plot which receives the event and performs an action.}

\item{action}{the action to be performed at target plot using input information.}

\item{flag}{indicates special flags for custom features.}
}
\value{
a input object containing the input field information.
}
\description{
Helper function for writing input fields in interactive Glimma plots
}
\examples{

}

