% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glBar.R
\name{glBar.default}
\alias{glBar.default}
\title{Glimma MD Plot}
\usage{
\method{glBar}{default}(x, yval, names.arg = rownames(x), ndigits = NULL,
  signif = 6, xlab = NULL, ylab = yval, main = NULL, height = 400,
  width = 500, colval = NULL, annot = yval, flag = NULL, info = NULL,
  ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{yval}{the column name for the x-axis values.}

\item{names.arg}{the column name for the label on each bar.}

\item{ndigits}{the number of digits after the decimal to round to in the tooltip (overrides signif).}

\item{signif}{the number of significant figures to display in the tooltip.}

\item{xlab}{the label on the x-axis.}

\item{ylab}{the label on the y-axis.}

\item{main}{the title for the plot.}

\item{height}{the height of the plot (in pixels).}

\item{width}{the width of the plot (in pixels).}

\item{colval}{the colours for each data point.}

\item{annot}{the columns to display in the tooltip.}

\item{flag}{the special flag to indicate special plot.}

\item{info}{additional information for plotting.}

\item{...}{additional arguments.}
}
\value{
A chart object containing the information to create an interactive bar plot.
}
\description{
Default method for interactive bar plot.
}
\examples{
data(mtcars)
counts <- table(mtcars$gear)
data <- data.frame(nGears=as.numeric(names(counts)), Count=as.numeric(counts))
\donttest{
plot1 <- glBar(data, "Count", "nGears", ylab="Number of Gears")
glimma(plot1, layout=c(1,1), launch=TRUE)
}
}
\author{
Shian Su
}
