% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimma.R
\name{glimma}
\alias{glimma}
\title{Core glimma plot manager. Generates environment for glimma plots.}
\usage{
glimma(..., layout = c(1, 1))
}
\arguments{
\item{...}{chart or link objects}

\item{layout}{numeric vector specifying rows and columns of plots to create.}
}
\value{
Generates interactive plots based on filling layout row by row from left to right.
}
\description{
Core glimma plot manager. Generates environment for glimma plots.
}
\examples{

}

