% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{gllink}
\alias{gllink}
\title{Plot linkages}
\usage{
gllink(from, to, src = "none", dest = "none", flag = "none",
  both = FALSE, info = "none")
}
\arguments{
\item{from}{the index of the plot from which the event is dispatched.}

\item{to}{the index of the plot which receives the event and performs an action.}

\item{src}{the action that is performed in the "from" plot.}

\item{dest}{the action that is performed in the "to" plot.}

\item{flag}{indicates special links for particular chart types.}

\item{both}{creates symmetric links whereby the "dest" action in "to" also triggers the "src" action in "from".}

\item{info}{additional info for creating the link.}
}
\value{
a link object containing the plot linking information.
}
\description{
Helper function for writing the link properties in interactive Glimma plots
}
\examples{

}

