% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDPlot.default}
\alias{glMDPlot.default}
\title{Glimma MD Plot}
\usage{
\method{glMDPlot}{default}(x, xval, yval, counts, anno, groups, samples,
  status = rep(0, nrow(x)), xlab = xval, ylab = yval,
  search.by = "Symbols", jitter = 30, display.columns = c("GeneID"),
  id.column = "GeneID", cols = c("#0000FF", "#858585", "#B32222"),
  path = getwd(), folder = "glimma-plots", html = "MD-Plot",
  launch = TRUE, ...)
}
\arguments{
\item{x}{the DGELRT object.}

\item{xval}{the column to plot on x axis.}

\item{yval}{the column to plot on y axis.}

\item{counts}{the matrix containing all counts.}

\item{anno}{the data.frame containing gene annotations.}

\item{groups}{the factor containing experimental groups of the samples.}

\item{samples}{the names of the samples.}

\item{status}{vector giving the control status of data point, of same length as the number of rows of object. If NULL, then all points are plotted in the default colour.}

\item{xlab}{the label on the x axis.}

\item{ylab}{the label on the y axis.}

\item{search.by}{the name of the column which will be used to search for data points. (should contain unique values)}

\item{jitter}{the amount of jitter to apply to the samples in the expressions plot.}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips.}

\item{id.column}{the column containing unique identifiers for each gene.}

\item{cols}{vector of strings denoting colours corresponding to control status -1, 0 and 1. (may be R named colours or Hex values)}

\item{path}{the path in which the folder will be created.}

\item{folder}{the name of the fold to save html file to.}

\item{html}{the name of the html file to save plots to.}

\item{launch}{TRUE to launch plot after call.}

\item{...}{additional arguments to be passed onto the MD plot.}
}
\value{
Draws a two-panel interactive MD plot in an html page. The left plot shows the log-fold-change vs average expression. The right plot shows the expression levels of a particular gene of each sample.
}
\description{
Draw an interactive MD plot from a DGELRT object
}
\author{
Shian Su
}

