% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDPlot.R
\name{glMDPlot.DESeqDataSet}
\alias{glMDPlot.DESeqDataSet}
\title{Glimma MD Plot}
\usage{
\method{glMDPlot}{DESeqDataSet}(x, counts = NULL, anno, groups, samples,
  status = rep(0, nrow(x)), transform = FALSE, side.xlab = "Group",
  side.ylab = "logMean", side.log = FALSE,
  side.gridstep = ifelse(!transform || side.log, FALSE, 0.5), jitter = 30,
  side.main = "GeneID", display.columns = NULL, cols = c("#00bfff",
  "#858585", "#ff3030"), sample.cols = rep("#1f77b4", ncol(x)),
  path = getwd(), folder = "glimma-plots", html = "MD-Plot",
  launch = TRUE, ...)
}
\arguments{
\item{x}{the DESeqDataSet object.}

\item{...}{additional arguments to be passed onto the MD plot. (main, xlab,
ylab can be set for the left plot)}
}
\value{
Draws a two-panel interactive MD plot in an html page. The left plot
  shows the log-fold-change vs average expression. The right plot shows the
  expression levels of a particular gene of each sample. Hovering over points
  on left plot will plot expression level for corresponding gene, clicking on
  points will fix the expression plot to gene. Clicking on rows on the table
  has the same effect as clicking on the corresponding gene in the plot.
}
\description{
Draw an interactive MD plot from a DESeqDataSet object
}
\author{
Shian Su
}

