% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDPlot.DGEExact}
\alias{glMDPlot.DGEExact}
\title{Draw an interactive MD plot from a DGELRT objet}
\usage{
\method{glMDPlot}{DGEExact}(x, counts, anno, groups, samples, status = rep(0,
  nrow(x)), coef = ncol(x$coefficients), p.adj.method = "BH",
  search.by = "Symbols", display.columns = c("GeneID"),
  cols = c("#858585", "#B32222", "#0000FF"), ...)
}
\arguments{
\item{x}{the DGEExact object.}

\item{counts}{the matrix containing all counts.}

\item{anno}{the data.frame containing gene annotations.}

\item{samples}{the names of the samples.}

\item{group}{the factor containing experimental groups of the samples.}
}
\value{
writes an interactive MD plot
}
\description{
Draw an interactive MD plot from a DGELRT objet
}
\examples{

}

