% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDPlot.DGELRT}
\alias{glMDPlot.DGELRT}
\title{Draw an interactive MD plot}
\usage{
\method{glMDPlot}{DGELRT}(x, counts, anno, groups, samples, status = rep(0,
  nrow(x)), coef = ncol(x$coefficients), p.adj.method = "BH",
  search.by = "Symbol", display.cols = c("GeneID"), ...)
}
\arguments{
\item{x}{the DGELRT object.}

\item{counts}{the matrix containing all counts.}

\item{anno}{the data.frame containing gene annotations.}

\item{samples}{the names of the samples.}

\item{group}{the factor containing experimental groups of the samples.}
}
\value{
writes an interactive MD plot
}
\description{
Draw an interactive MD plot
}
\examples{

}

