% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDPlot.hidden}
\alias{glMDPlot.hidden}
\title{Draw an interactive MD plot}
\usage{
\method{glMDPlot}{hidden}(plotting.data, sample.exp, display.columns, ...)
}
\description{
Draw an interactive MD plot
}
\examples{

}

