% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDPlot}
\alias{glMDPlot}
\title{Glimma MD Plot}
\usage{
glMDPlot(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{...}{additional arguments affecting the plots produced. See specific methods for detailed arguments.}
}
\description{
Draw an interactive MD plot
}
\examples{

}
\author{
Shian Su
}
\seealso{
\code{\link{glMDPlot.default}}, \code{\link{glMDPlot.DGELRT}}, \code{\link{glMDPlot.DGEExact}}, \code{\link{glMDPlot.MArrayLM}}, \code{\link{glMDPlot.DESeqDataSet}}
}

