% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDUtils.R
\name{glMDRmd}
\alias{glMDRmd}
\title{glMDPlot Rmarkdown link and instructions}
\usage{
glMDRmd(html = "MD-Plot")
}
\arguments{
\item{html}{name of the HTML page containing plots from glMDPlot.}
}
\value{
None
}
\description{
When run inside of a text-block of Rmarkdown document using `r ...` this
produces a link and instructions about the usage of the interactive plots.
}
\examples{
glMDRmd()

}
\seealso{
\code{\link{glMDPlot}}
}
