% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDSPlot.DESeqResults}
\alias{glMDSPlot.DESeqResults}
\title{Draw an interactive MD plot from a DESeqResults object}
\usage{
\method{glMDSPlot}{DESeqResults}(x, anno)
}
\arguments{
\item{x}{the DGEExact object.}

\item{anno}{the data.frame containing gene annotations.}

\item{counts}{the matrix containing all counts.}

\item{group}{the factor containing experimental groups of the samples.}

\item{samples}{the names of the samples.}
}
\value{
writes an interactive MD plot
}
\description{
Draw an interactive MD plot from a DESeqResults object
}
\examples{

}

