% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDSPlot.DGELRT}
\alias{glMDSPlot.DGELRT}
\title{Draw an interactive MD plot}
\usage{
\method{glMDSPlot}{DGELRT}(x, counts, anno, group, samples)
}
\arguments{
\item{x}{the DGELRT object.}

\item{counts}{the matrix containing all counts.}

\item{anno}{the data.frame containing gene annotations.}
}
\value{
writes an interactive MD plot
}
\description{
Draw an interactive MD plot
}
\examples{

}

