% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDSPlot.R
\name{glMDSPlot}
\alias{glMDSPlot}
\title{Create an interactive bar plot object}
\usage{
glMDSPlot(x, ...)
}
\arguments{
\item{x}{A data.frame containing data to plot}
}
\value{
a "chart" object containing information used by glimma to create plots
}
\description{
Create an interactive bar plot object
}
\examples{

}

