% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glScatter.R
\name{glScatter}
\alias{glScatter}
\title{Create an interactive scatter plot object}
\usage{
glScatter(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{...}{additional arguments depending on input object type.}

\item{xval}{the column name for the x-axis values.}

\item{yval}{the column name for the y-axis values.}

\item{id}{the column name for unique identifiers.}

\item{ndigits}{the number of digits after the decimal to round to in the tooltip (overrides signif).}

\item{signif}{the number of significant figures to display in the tooltip.}

\item{xlab}{the label on the x-axis.}

\item{ylab}{the label on the y-axis.}

\item{main}{the title for the plot.}

\item{height}{the height of the plot (in pixels).}

\item{width}{the width of the plot (in pixels).}

\item{colval}{the colours for each data point.}

\item{annot}{the columns to display in the tooltip.}
}
\value{
A chart object containing the information to create an interactive scatter plot.
}
\description{
Create an interactive scatter plot object
}
\examples{
data(iris)
plot1 <- glScatter(iris, xval="Sepal.Length", yval="Sepal.Width", colval="Species")
glimma(plot1, c(1,1))

}

