% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glTable.R
\name{glTable}
\alias{glTable}
\title{Glimma Table}
\usage{
glTable(target, columns)
}
\arguments{
\item{target}{the index of the plot which receives the event and performs an action.}

\item{action}{the action to be performed at target plot using input information.}

\item{idval}{the column from which the autocomplete list will be populated.}

\item{flag}{indicates special flags for custom features.}
}
\value{
a input object containing the input field information.
}
\description{
Create a table using the data from a chart.
}

