% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gltablink.R
\name{gltablink}
\alias{gltablink}
\title{Plot linkages}
\usage{
gltablink(from, to, action = "none", info = "none")
}
\arguments{
\item{from}{the index of the source table.}

\item{to}{the index of the plot which receives the event and performs an action.}

\item{action}{the action that is performed in the plot.}

\item{info}{additional info for creating the link.}
}
\value{
a link object containing the plot linking information.
}
\description{
Helper function for writing the link properties in interactive Glimma plots
}
