% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hasColumns}
\alias{hasColumns}
\title{Column checker}
\usage{
hasColumns(df, columns)
}
\arguments{
\item{df}{the data frame to check.}

\item{columns}{the columns that should exist in the data frame.}
}
\value{
stops program with an error if column cannot be found in df
}
\description{
Check if data.frame controls all the listed columns
}

