% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexcol.R
\name{is.hex}
\alias{is.hex}
\title{Hexcode colours}
\usage{
is.hex(x)
}
\arguments{
\item{x}{the colour value(s) to check.}
}
\value{
Logical vector indicating if strings(s) are valid hex representations
}
\description{
Check if string(s) are valid hex colour representation
}
