% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{lymphomaRNAseq}
\alias{lymphomaRNAseq}
\title{Mouse based RNAseq data for study of smchd1 gene.}
\description{
Mouse based RNAseq data for study of smchd1 gene.
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S2213596015301306}
}
\author{
Ruijie Liu, Kelan Chen, Natasha Jansz, Marnie E. Blewitt, Matthew E. Ritchie
}
\keyword{RNAseq}
