% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeChartJson}
\alias{makeChartJson}
\title{JSON converter for chart objects}
\usage{
makeChartJson(chart)
}
\arguments{
\item{chart}{the chart object to be converted into JSON}
}
\value{
a stringified JSON object containing the chart data.
}
\description{
Function to make json object from a chart, ignoring the json property
}

