% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonConvert.R
\name{makeJson}
\alias{makeJson}
\title{JSON converter for R objects}
\usage{
makeJson(x, ...)
}
\arguments{
\item{x}{the object to be converted into JSON}

\item{...}{additional arguments}
}
\value{
a stringified JSON object.
}
\description{
Function to generate json strings from
}
