% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{NumToHexCol}
\alias{NumToHexCol}
\title{Numeric to hex colour converter}
\usage{
NumToHexCol(x)
}
\arguments{
\item{x}{the colour value(s) to be converted to hex values.}
}
\value{
hex codes for colours
}
\description{
Functions to convert numbers into corresponding hex codes for colours
}

