% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexcol.R
\name{as.hexcol}
\alias{as.hexcol}
\title{Numeric to hex colour converter}
\usage{
as.hexcol(x)
}
\arguments{
\item{x}{the colour value(s) to be converted to hex values.}
}
\value{
hex codes for colours
}
\description{
Convert numbers and R colour strings into corresponding hex codes for colours
}
\examples{
as.hexcol(c(1, 2, 3))
as.hexcol(c("red", "black", "green"))

}
