% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glBar.R
\name{glBar}
\alias{glBar}
\title{Glimma MD Plot}
\usage{
glBar(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{...}{additional arguments depending on input object type.}
}
\value{
A chart object containing the information to create an interactive bar plot.
}
\description{
Create an interactive bar plot object.
}
\examples{
data(mtcars)
counts <- table(mtcars$gear)
data <- data.frame(nGears=as.numeric(names(counts)), Count=as.numeric(counts))
\donttest{
plot1 <- glBar(data, "Count", "nGears", ylab="Number of Gears")
glimma(plot1, layout=c(1,1), launch=TRUE)
}
}
\seealso{
\code{\link{glBar.default}}
}
\author{
Shian Su
}
