% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimma.R
\name{glimma}
\alias{glimma}
\title{Core glimma plot manager. Generates environment for glimma plots.}
\usage{
glimma(..., layout = c(1, 1), folder = "glimma", overwrite = FALSE)
}
\arguments{
\item{...}{the jschart or jslink objects for processing.}

\item{layout}{the numeric vector representing the number of rows and columns in plot window.}

\item{folder}{the name of the folder to save plots to.}

\item{overwrite}{the option to overwrite existing folder if it already exists.}
}
\value{
Generates interactive plots based on filling layout row by row from left to right.
}
\description{
Core glimma plot manager. Generates environment for glimma plots.
}
\examples{
data(iris)
plot1 <- glScatter(iris, xval="Sepal.Length", yval="Sepal.Width", colval="Species")
glimma(plot1, c(1,1))

}

