% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{glimma}
\alias{glimma}
\title{Glimma: interactive graphics from limma}
\description{
The Glimma package provides iteractive versions of plots frequently used in
the limma package. Currently the MDS and MD plots have been implemented. The
functions can be used with both limma, edgeR and DESeq objecs.
}
\section{Main functions}{

\code{\link{glMDSPlot}},
\code{\link{glMDPlot}},
\code{\link{glXYPlot}}
}

