% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimma.R
\name{glimma}
\alias{glimma}
\title{Glimma plot manager}
\usage{
glimma(..., layout = c(1, 1), path = getwd(), folder = "glimma-plots",
  html = "index", overwrite = TRUE, launch = TRUE)
}
\arguments{
\item{...}{the jschart or jslink objects for processing.}

\item{layout}{the numeric vector representing the number of rows and columns in plot window.}

\item{path}{the path in which the folder will be created.}

\item{folder}{the name of the fold to save html file to.}

\item{html}{the name of the html file to save plots to.}

\item{overwrite}{the option to overwrite existing folder if it already exists.}

\item{launch}{TRUE to launch plot after call.}
}
\value{
Generates interactive plots based on filling layout row by row from left to right.
}
\description{
Core glimma plot manager. Generates environment for glimma plots.
}
\examples{
data(iris)
\donttest{
plot1 <- glScatter(iris, xval="Sepal.Length", yval="Sepal.Width", colval="Species")
glimma(plot1, c(1,1))
}

}
