% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gllink.R
\name{gllink}
\alias{gllink}
\title{Plot linkages}
\usage{
gllink(from, to, src = "none", dest = "none", flag = "none",
  both = FALSE, info = "none")
}
\arguments{
\item{from}{the index of the plot from which the event is dispatched.}

\item{to}{the index of the plot which receives the event and performs an action.}

\item{src}{the action that is performed in the "from" plot.}

\item{dest}{the action that is performed in the "to" plot.}

\item{flag}{indicates special links for particular chart types.}

\item{both}{creates symmetric links whereby the "dest" action in "to" also triggers the "src" action in "from".}

\item{info}{additional info for creating the link.}
}
\value{
a link object containing the plot linking information.
}
\description{
Helper function for writing the link properties in interactive Glimma plots
}
\examples{
data(iris)
data <- data.frame(Name=paste("Flower", 1:nrow(iris), sep="-"), iris)
\donttest{
plot1 <- glScatter(data, xval="Sepal.Length", yval="Sepal.Width", colval="Species")
plot2 <- glScatter(data, xval="Species", yval="Petal.Length", colval="Species")
link1 <- gllink(1, 2, src="hover", dest="hover", both=TRUE)
glimma(plot1, plot2, link1, layout=c(1,2))
}
}
