% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDPlot.R
\name{glMDPlot.DESeqDataSet}
\alias{glMDPlot.DESeqDataSet}
\title{Glimma MD Plot}
\usage{
\method{glMDPlot}{DESeqDataSet}(x, anno, groups, samples, status = rep(0,
  nrow(x)), search.by = "Symbols", jitter = 30, id.column = "GeneID",
  display.columns = id.column, cols = c("#0000FF", "#858585", "#B32222"),
  sample.cols = rep("#1f77b4", ncol(x)), table = TRUE, path = getwd(),
  folder = "glimma-plots", html = "MD-Plot", launch = TRUE, ...)
}
\arguments{
\item{x}{the DESeqDataSet object.}

\item{anno}{the data.frame containing gene annotations.}

\item{groups}{the factor containing experimental groups of the samples.}

\item{samples}{the names of the samples.}

\item{status}{vector giving the control status of data point, of same length as the number of rows of object. If NULL, then all points are plotted in the default colour.}

\item{search.by}{the name of the column which will be used to search for data points. (should contain unique values)}

\item{jitter}{the amount of jitter to apply to the samples in the expressions plot.}

\item{id.column}{the column containing unique identifiers for each gene.}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips and table.}

\item{cols}{vector of strings denoting colours corresponding to control status -1, 0 and 1. (may be R named colours or Hex values)}

\item{sample.cols}{vector of strings denoting colours for each sample point on the expression plot.}

\item{table}{logical variable for whether a table of the data should appear on the bottom of the HTML page.}

\item{path}{the path in which the folder will be created.}

\item{folder}{the name of the fold to save html file to.}

\item{html}{the name of the html file to save plots to.}

\item{launch}{TRUE to launch plot after call.}

\item{...}{additional arguments to be passed onto the MD plot. (main, xlab, ylab can be set for the left plot)}
}
\value{
Draws a two-panel interactive MD plot in an html page. The left plot
shows the log-fold-change vs average expression. The right plot shows the 
expression levels of a particular gene of each sample. Hovering over points 
on left plot will plot expression level for corresponding gene, clicking 
on points will fix the expression plot to gene. Clicking on rows on the table
has the same effect as clicking on the corresponding gene in the plot.
}
\description{
Draw an interactive MD plot from a DESeqDataSet object
}
\author{
Shian Su
}

