% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R
\name{glMDPlot.DGEExact}
\alias{glMDPlot.DGEExact}
\title{Draw an interactive MD plot from a DGELRT objet}
\usage{
\method{glMDPlot}{DGEExact}(x, counts, anno, groups, samples, status = rep(0,
  nrow(x)), coef = ncol(x$coefficients), p.adj.method = "BH",
  search.by = "Symbols", display.columns = c("GeneID"),
  id.column = "GeneID", cols = c("#0000FF", "#858585", "#B32222"),
  path = getwd(), folder = "glimma-plots", html = "MD-Plot",
  launch = TRUE, ...)
}
\arguments{
\item{x}{the DGEExact object.}

\item{counts}{the matrix containing all counts.}

\item{anno}{the data.frame containing gene annotations.}

\item{samples}{the names of the samples.}

\item{status}{vector giving the control status of data point, of same length as the number of rows of object. If NULL, then all points are plotted in the default color.}

\item{coef}{integer or character index vector indicating which column of object to plot.}

\item{p.adj.method}{character vector indicating multiple testing correction method. (defauls to "BH")}

\item{search.by}{the name of the column which will be used to search for data points. (should contain unique values)}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips.}

\item{cols}{vector of strings denoting colours corresponding to control status -1, 0 and 1. (may be R named colours or Hex values)}

\item{launch}{TRUE to launch plot after call.}

\item{group}{the factor containing experimental groups of the samples.}
}
\description{
Draw an interactive MD plot from a DGELRT objet
}
\examples{

}

