% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDPlot.R
\name{glMDPlot.DGELRT}
\alias{glMDPlot.DGELRT}
\title{Glimma MD Plot}
\usage{
\method{glMDPlot}{DGELRT}(x, counts = NULL, anno = NULL, groups = rep(0,
  ncol(x)), samples = NULL, status = rep(0, nrow(x)), transform = TRUE,
  side.xlab = "Group", side.ylab = "logCPM", side.log = FALSE,
  side.gridstep = ifelse(!transform || side.log, FALSE, 0.5),
  p.adj.method = "BH", search.by = "Symbols", jitter = 30,
  id.column = "GeneID", display.columns = NULL, cols = c("#0000FF",
  "#858585", "#B32222"), sample.cols = rep("#1f77b4", ncol(counts)),
  table = TRUE, path = getwd(), folder = "glimma-plots",
  html = "MD-Plot", launch = TRUE, ...)
}
\arguments{
\item{x}{the DGELRT object.}

\item{counts}{the matrix containing all counts.}

\item{anno}{the data.frame containing gene annotations.}

\item{groups}{the factor containing experimental groups of the samples.}

\item{samples}{the names of the samples.}

\item{status}{vector giving the control status of data point, of same length as the number of rows of object. If NULL, then all points are plotted in the default colour.}

\item{transform}{TRUE if counts are raw and should be cpm transformed, FALSE if counts are already transformed to expression scale.}

\item{side.xlab}{label for x axis on right side plot.}

\item{side.ylab}{label for y axis on right side plot.}

\item{side.log}{TRUE to plot expression on the side plot on log scale.}

\item{side.gridstep}{intervals along which to place grid lines on y axis. Currently only available for linear scale.}

\item{p.adj.method}{character vector indicating multiple testing correction method. See \code{\link{p.adjust}} for available methods. (defaults to "BH")}

\item{search.by}{the name of the column which will be used to search for data points. (should contain unique values)}

\item{jitter}{the amount of jitter to apply to the samples in the expressions plot.}

\item{id.column}{the column containing unique identifiers for each gene.}

\item{display.columns}{character vector containing names of columns to display in mouseover tooltips and table.}

\item{cols}{vector of strings denoting colours corresponding to control status -1, 0 and 1. (may be R named colours or Hex values)}

\item{sample.cols}{vector of strings denoting colours for each sample point on the expression plot.}

\item{table}{logical variable for whether a table of the data should appear on the bottom of the HTML page.}

\item{path}{the path in which the folder will be created.}

\item{folder}{the name of the fold to save html file to.}

\item{html}{the name of the html file to save plots to.}

\item{launch}{TRUE to launch plot after call.}

\item{...}{additional arguments to be passed onto the MD plot. (main, xlab, ylab can be set for the left plot)}
}
\value{
Draws a two-panel interactive MD plot in an html page. The left plot
shows the log-fold-change vs average expression. The right plot shows the
expression levels of a particular gene of each sample. Hovering over points
on left plot will plot expression level for corresponding gene, clicking
on points will fix the expression plot to gene. Clicking on rows on the table
has the same effect as clicking on the corresponding gene in the plot.
}
\description{
Draw an interactive MD plot from a DGELRT object
}
\author{
Shian Su
}

