% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDPlot.R
\name{glMDPlot}
\alias{glMDPlot}
\title{Glimma MD Plot}
\usage{
glMDPlot(x, ...)
}
\arguments{
\item{x}{the DE object to plot.}

\item{...}{additional arguments affecting the plots produced. See specific
methods for detailed arguments.}
}
\value{
Draws a two-panel interactive MD plot in an html page. The left plot
  shows the log-fold-change vs average expression. The right plot shows the
  expression levels of a particular gene of each sample. Hovering over points
  on left plot will plot expression level for corresponding gene, clicking on
  points will fix the expression plot to gene. Clicking on rows on the table
  has the same effect as clicking on the corresponding gene in the plot.
}
\description{
Draw an interactive MD plot
}
\examples{
library(limma)
library(edgeR)

data(lymphomaRNAseq)
x <- lymphomaRNAseq

sel <- rowSums(cpm(x$counts)>0.5)>=3
x <- x[sel,]

genotype <- relevel(x$samples$group, "Smchd1-null")
x <- calcNormFactors(x, method="TMM")
des <- model.matrix(~genotype)

## Apply voom with sample quality weights and fit linear model
v <- voomWithQualityWeights(x, design=des, normalization="none", plot=FALSE)
vfit <- lmFit(v,des)

## Apply treat relative to a fold-change of 1.5
vtfit <- treat(vfit,lfc=log2(1.5))
vfit <- eBayes(vfit)
results <- decideTests(vfit,p.value=0.01)

\donttest{
glMDPlot(vfit, counts=x$counts, anno=x$genes, groups=genotype, samples=1:7,
         status=results[,2], main="MD plot: Wild-type vs Smchd1",
         display.columns=c("Symbols", "GeneID", "GeneName"),
         folder="Smchd1-Lymphoma")
}

}
\seealso{
\code{\link{glMDPlot.default}}, \code{\link{glMDPlot.DGELRT}},
  \code{\link{glMDPlot.DGEExact}}, \code{\link{glMDPlot.MArrayLM}},
  \code{\link{glMDPlot.DESeqDataSet}}
}
\author{
Shian Su
}
