% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDSPlot.R
\name{glMDSPlot.DGEList}
\alias{glMDSPlot.DGEList}
\title{Draw an interactive MD plot from a DGEList object}
\usage{
\method{glMDSPlot}{DGEList}(x, top = 500, labels = 1:ncol(x),
  gene.selection = "pairwise", main = "MDS Plot")
}
\description{
Draw an interactive MD plot from a DGEList object
}

