% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDSPlot.R
\name{glMDSPlot.DGEList}
\alias{glMDSPlot.DGEList}
\title{Draw an interactive MD plot from a DGEList object}
\usage{
\method{glMDSPlot}{DGEList}(x, top = 500, labels = 1:ncol(x),
  groups = rep(1, ncol(x)), gene.selection = "pairwise",
  main = "MDS Plot", launch = TRUE, ...)
}
\arguments{
\item{x}{the object containing the gene expressions.}

\item{top}{the number of top most variable genes to use.}

\item{labels}{the labels for each sample.}

\item{groups}{the experimental group to which samples belong.}

\item{gene.selection}{"pairwise" if most variable genes are to be chosen for each pair of samples or "common" to select the same genes for all comparisons.}

\item{main}{the title of the plot.}

\item{launch}{TRUE to launch plot after call.}

\item{...}{additional arguments.}
}
\description{
Draw an interactive MD plot from a DGEList object
}

