% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDSPlot.R
\name{glMDSPlot}
\alias{glMDSPlot}
\title{Create an interactive bar plot object}
\usage{
glMDSPlot(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{top}{the number of top genes used to calculate distances.}

\item{labels}{the haracter vector of sample names or labels. Defaults to colnames(x).}
}
\value{
a "chart" object containing information used by glimma to create plots.
}
\description{
Create an interactive bar plot object
}
\examples{

}

