% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDSPlot.R
\name{glMDSPlot}
\alias{glMDSPlot}
\title{Draw an interactive MDS plot}
\usage{
glMDSPlot(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{...}{additional arguments affecting the plots produced. See specific methods for detailed arguments.}
}
\value{
writes an interactive MDS plot
}
\description{
Draw an interactive MDS plot
}
\examples{

}
\seealso{
\code{\link{glMDSPlot.default}}, \code{\link{glMDSPlot.DGEList}}
}

