% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMD.R, R/glMDSPlot.R
\name{glMDSPlot}
\alias{glMDSPlot}
\title{Darw an interactive MD plot}
\usage{
glMDSPlot(x, ...)

glMDSPlot(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{x}{the data.frame containing data to plot.}

\item{top}{the number of top genes used to calculate distances.}

\item{labels}{the haracter vector of sample names or labels. Defaults to colnames(x).}

\item{gene.selection}{the method for calculating distances.}
}
\value{
writes an interactive MDS plot

writes an interactive MDS plot
}
\description{
Darw an interactive MD plot

Draw an interactive MDS plot
}
\examples{


}

