% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glMDSPlot.R
\name{glMDSPlot}
\alias{glMDSPlot}
\title{Glimma MDS Plot}
\usage{
glMDSPlot(x, ...)
}
\arguments{
\item{x}{the matrix containing the gene expressions.}

\item{...}{additional arguments.}
}
\value{
Draws a two-panel interactive MDS plot in an html page. The left
  panel contains the plot between two MDS dimensions, with annotations
  displayed on hover. The right panel contains a bar plot of the eigenvalues
  of each dimension, clicking on any of the bars will plot the corresponding
  dimension against the next dimension.
}
\description{
Glimma MDS Plot

Draw an interactive MD plot from a DGEList object with distances calculated
from most variable genes.
}
\examples{
data(lymphomaRNAseq)
genotype <- relevel(lymphomaRNAseq$samples$group, "Smchd1-null")
\donttest{
glMDSPlot(lymphomaRNAseq, labels=1:7, groups=genotype)
}

}
\seealso{
\code{\link{glMDSPlot.default}}, \code{\link{glMDSPlot.DGEList}}
}
\author{
Shian Su, Gordon Smyth
}
