% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glScatter.R
\name{glScatter}
\alias{glScatter}
\title{Glimma Scatter Plot}
\usage{
glScatter(x, ...)
}
\arguments{
\item{x}{the data.frame containing data to plot.}

\item{...}{additional arguments depending on input object type.}
}
\value{
A chart object containing the information to create an interactive scatter plot.
}
\description{
Create an interactive scatter plot object
}
\examples{
data(iris)
\donttest{
plot1 <- glScatter(iris, xval="Sepal.Length", yval="Sepal.Width", colval="Species")
glimma(plot1, c(1,1))
}
}
\author{
Shian Su
}
