% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glTable.R
\name{glTable}
\alias{glTable}
\title{Glimma Table}
\usage{
glTable(target, columns)
}
\arguments{
\item{target}{the index of the plot from which data is drawn.}

\item{columns}{the columns of data to plot.}
}
\value{
a input object containing the input field information.
}
\description{
Create a table using the data from a chart.
}
