% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonConvert.R
\name{makeJson.data.frame}
\alias{makeJson.data.frame}
\title{JSON converter for data frames}
\usage{
\method{makeJson}{data.frame}(df, convert.logical = TRUE,
  dataframe = c("rows", "columns"))
}
\arguments{
\item{df}{the data.frame to be converted into JSON}

\item{convert.logical}{whether to convert logicals into strings "TRUE" and "FALSE"}

\item{dataframe}{how to encode data.frame objects: must be one of 'rows', 'columns'}
}
\value{
a stringified JSON, the data.frame is encoded as a vector of objects,
with each column being one object with keys corresponding to column names.
}
\description{
Function to create a JSON from a data.frame
}
